+++ pysatNASA/instruments/methods/general.py
@@ -67,8 +67,16 @@ def clean(self):
         coords = [key for key in self.data.coords.keys()]
 
     for key in self.variables:
+        # Check for symmetric dims
+        # Indicates transformation matrix, xarray cannot broadcast
+        if self.pandas_format:
+            # True by default
+            unique_dims = True
+        else:
+            # Check for multiple dims
+            unique_dims = len(self[key].dims) == len(np.unique(self[key].dims))
         # Skip over the coordinates when cleaning
-        if key not in coords:
+        if key not in coords and unique_dims:
             fill = self.meta[key, self.meta.labels.fill_val]
 
             # Replace fill with nan