+++ CHANGELOG.md
@@ -3,10 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.X.X] - 2023-XX-XX
-* Bug Fix
-  * New window needs to be integer for calculate_imf_steadiness
+* New Instruments
+  * MAVEN mag
+  * MAVEN SEP
+  * MAVEN in situ
 * Documentation
   * Added example of how to export data for archival
+* Bug Fixes
+  * New window needs to be integer for calculate_imf_steadiness
+  * Fix general clean routine to skip transformation matrices
 * Maintenance
   * Implemented unit tests for cleaning warnings
   * Use pip install for readthedocs